+++ ui/main/game/galactic_war/cards/gwc_combat_structures.js
@@ -18,13 +18,13 @@ define([
     getContext: gwoCard.getContext,
     deal: function (system, context) {
       var chance = 28;
-      var dist = system.distance();
       if (
-        (context.totalSize <= GW.balance.numberOfSystems[0] && dist > 4) ||
-        (context.totalSize <= GW.balance.numberOfSystems[1] && dist > 6) ||
-        (context.totalSize <= GW.balance.numberOfSystems[2] && dist > 9) ||
-        (context.totalSize <= GW.balance.numberOfSystems[3] && dist > 11) ||
-        dist > 13
+        gwoCard.farForSize(
+          system,
+          context,
+          GW.balance.numberOfSystems,
+          gwoCard.farLadders.spike
+        )
       ) {
         chance = 142;
       }