+++ ui/main/game/galactic_war/cards/gwc_damage_commander.js
@@ -1,7 +1,8 @@
 define([
   "shared/gw_common",
+  "coui://ui/mods/com.pa.quitch.gwaioverhaul/shared/cards.js",
   "coui://ui/mods/com.pa.quitch.gwaioverhaul/shared/unit_groups.js",
-], function (GW, gwoGroup) {
+], function (GW, gwoCard, gwoGroup) {
   return {
     visible: _.constant(true),
     describe: _.constant(
@@ -21,16 +22,18 @@ define([
     },
     deal: function (system, context) {
       var chance = 24;
-      var dist = system.distance();
       if (
         context.totalSize <= GW.balance.numberOfSystems[0] ||
         context.totalSize <= GW.balance.numberOfSystems[1]
       ) {
         chance = 12;
       } else if (
-        (context.totalSize <= GW.balance.numberOfSystems[2] && dist > 6) ||
-        (context.totalSize <= GW.balance.numberOfSystems[3] && dist > 9) ||
-        dist > 12
+        gwoCard.farForSize(
+          system,
+          context,
+          GW.balance.numberOfSystems,
+          gwoCard.farLadders.moderate
+        )
       ) {
         chance = 120;
       }