+++ ui/main/game/galactic_war/cards/gwc_speed_orbital.js
@@ -17,13 +17,15 @@ define([
     }),
     getContext: gwoCard.getContext,
     deal: function (system, context) {
-      var dist = system.distance();
       var chance = context.totalSize <= GW.balance.numberOfSystems[1] ? 16 : 32;
 
       if (
-        (context.totalSize <= GW.balance.numberOfSystems[2] && dist > 6) ||
-        (context.totalSize <= GW.balance.numberOfSystems[3] && dist > 9) ||
-        (context.totalSize > GW.balance.numberOfSystems[3] && dist > 12)
+        gwoCard.farForSize(
+          system,
+          context,
+          GW.balance.numberOfSystems,
+          gwoCard.farLadders.moderate
+        )
       ) {
         chance = 166;
       }